update custom resource lambda to not update ES (handled by cfn)
