+++ cfn-custom-resource/es-cognito/requirements.txt
@@ -1,0 +1,1 @@
+cfnresponse
